requirements.txt
numpy
pandas
pillow
transformers==4.37.2
timm
pytesseract
Flask